requirements.txt
pandas
numpy
sklearn
flask
gunicorn
opencv-python
tensorflow-cpu
keras
Jinja2
